requirement.txt
anyio==3.5.0
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
dlib==19.23.1
duplicity==0.8.12.0
entrypoints==0.3
executing==0.8.3
fasteners==0.14.1
fastjsonschema==2.15.3
filelock==3.6.0
Flask==2.1.1
fonttools==4.33.1
future==0.18.2
httplib2==0.14.0
idna==2.8
image==1.5.33
imageio==2.17.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
imutils==0.5.4
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
keyring==18.0.1
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
nbclassic==0.3.7
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.10.4
networkx==2.8
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.5.64
packaging==21.3
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.6.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
Pygments==2.12.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.8
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2019.3
PyWavelets==1.3.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scikit-image==0.19.2
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
systemd-python==234
terminado==0.13.3
tifffile==2022.4.8
tinycss2==1.1.1
tornado==6.1
traitlets==5.1.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.14.1
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.1
xkit==0.0.0
zipp==3.8.0
